+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/WebServerDelegate.kt
@@ -9,7 +9,7 @@ import com.projectcitybuild.pcbridge.http.models.pcb.PlayerBan
 import com.projectcitybuild.pcbridge.http.models.pcb.RemoteConfigVersion
 import com.projectcitybuild.pcbridge.http.models.pcb.Warp
 import com.projectcitybuild.pcbridge.paper.core.libs.logger.log
-import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.services.RemoteConfig
+import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.RemoteConfig
 import com.projectcitybuild.pcbridge.paper.core.support.spigot.SpigotEventBroadcaster
 import com.projectcitybuild.pcbridge.webserver.HttpServerDelegate
 import java.util.UUID
