+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/features/announcements/listeners/AnnouncementConfigListener.kt
@@ -1,7 +1,7 @@
 package com.projectcitybuild.pcbridge.paper.features.announcements.listeners
 
 import com.projectcitybuild.pcbridge.paper.core.libs.logger.log
-import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.events.RemoteConfigUpdatedEvent
+import com.projectcitybuild.pcbridge.paper.features.config.events.RemoteConfigUpdatedEvent
 import com.projectcitybuild.pcbridge.paper.features.announcements.actions.StartAnnouncementTimer
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
