+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/features/building/listeners/FrameItemRemoveListener.kt
@@ -1,34 +1,34 @@
-package com.projectcitybuild.pcbridge.paper.features.invisframes.listeners
+package com.projectcitybuild.pcbridge.paper.features.building.listeners
 
 import com.projectcitybuild.pcbridge.paper.core.support.spigot.SpigotNamespace
+import org.bukkit.entity.GlowItemFrame
 import org.bukkit.entity.ItemFrame
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
-import org.bukkit.event.hanging.HangingPlaceEvent
+import org.bukkit.event.entity.EntityDamageByEntityEvent
 import org.bukkit.persistence.PersistentDataType
 
-class FramePlaceListener(
+class FrameItemRemoveListener(
     private val spigotNamespace: SpigotNamespace,
 ) : Listener {
     @EventHandler
-    fun onHangingPlaceEvent(event: HangingPlaceEvent) {
-        val isItemFrame = event.entity is ItemFrame
+    fun onEntityDamageByEntity(event: EntityDamageByEntityEvent) {
+        val entity = event.entity
+
+        val isItemFrame = entity is ItemFrame
         if (!isItemFrame) return
 
         val invisibleValue =
-            event.itemStack?.itemMeta?.persistentDataContainer?.getOrDefault(
+            entity.persistentDataContainer.getOrDefault(
                 spigotNamespace.invisibleKey,
                 PersistentDataType.BYTE,
                 0,
-            ) ?: 0.toByte()
-
+            )
         val isInvisibleFrame = invisibleValue == 1.toByte()
         if (isInvisibleFrame) {
-            event.entity.persistentDataContainer.set(
-                spigotNamespace.invisibleKey,
-                PersistentDataType.BYTE,
-                1,
-            )
+            val itemFrame = entity as ItemFrame
+            itemFrame.isVisible = true
+            itemFrame.isGlowing = entity is GlowItemFrame
         }
     }
 }
