+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/features/building/listeners/FramePlaceListener.kt
@@ -1,34 +1,34 @@
-package com.projectcitybuild.pcbridge.paper.features.invisframes.listeners
+package com.projectcitybuild.pcbridge.paper.features.building.listeners
 
 import com.projectcitybuild.pcbridge.paper.core.support.spigot.SpigotNamespace
-import org.bukkit.entity.GlowItemFrame
 import org.bukkit.entity.ItemFrame
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
-import org.bukkit.event.player.PlayerInteractEntityEvent
+import org.bukkit.event.hanging.HangingPlaceEvent
 import org.bukkit.persistence.PersistentDataType
 
-class FrameItemInsertListener(
+class FramePlaceListener(
     private val spigotNamespace: SpigotNamespace,
 ) : Listener {
     @EventHandler
-    fun onPlayerInteractEntityEvent(event: PlayerInteractEntityEvent) {
-        val entity = event.rightClicked
-
-        val isItemFrame = entity is ItemFrame
+    fun onHangingPlaceEvent(event: HangingPlaceEvent) {
+        val isItemFrame = event.entity is ItemFrame
         if (!isItemFrame) return
 
         val invisibleValue =
-            entity.persistentDataContainer.getOrDefault(
+            event.itemStack?.itemMeta?.persistentDataContainer?.getOrDefault(
                 spigotNamespace.invisibleKey,
                 PersistentDataType.BYTE,
                 0,
-            )
+            ) ?: 0.toByte()
+
         val isInvisibleFrame = invisibleValue == 1.toByte()
         if (isInvisibleFrame) {
-            val itemFrame = entity as ItemFrame
-            itemFrame.isVisible = false
-            itemFrame.isGlowing = entity is GlowItemFrame
+            event.entity.persistentDataContainer.set(
+                spigotNamespace.invisibleKey,
+                PersistentDataType.BYTE,
+                1,
+            )
         }
     }
 }
