+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/features/chat/repositories/ChatBadgeRepository.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.pcbridge.paper.features.chat.repositories
 
-import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.services.RemoteConfig
+import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.RemoteConfig
 import com.projectcitybuild.pcbridge.paper.architecture.store.Store
 import com.projectcitybuild.pcbridge.paper.features.chat.ChatBadgeFormatter
 import io.github.reactivecircus.cache4k.Cache
