+++ pcbridge-paper/src/main/kotlin/com/projectcitybuild/pcbridge/paper/features/joinmessages/listeners/ServerOverviewJoinListener.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.pcbridge.paper.features.joinmessages.listeners
 
-import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.services.RemoteConfig
+import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.RemoteConfig
 import net.kyori.adventure.text.minimessage.MiniMessage
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
