+++ pcbridge-paper/test/com/projectcitybuild/pcbridge/paper/features/staffchat/commands/ACommandTest.kt
@@ -1,7 +1,7 @@
 package com.projectcitybuild.pcbridge.paper.features.staffchat.commands
 
 import com.mojang.brigadier.context.CommandContext
-import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.services.RemoteConfig
+import com.projectcitybuild.pcbridge.paper.core.libs.remoteconfig.RemoteConfig
 import com.projectcitybuild.pcbridge.http.models.pcb.RemoteConfigKeyValues
 import com.projectcitybuild.pcbridge.http.models.pcb.RemoteConfigVersion
 import com.projectcitybuild.pcbridge.paper.PermissionNode
